backend/requirements.txt
# Core Web Framework
fastapi>=0.100.0,<0.115.0
uvicorn[standard]>=0.24.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0

# Database - MongoDB
motor>=3.3.0
pymongo>=4.6.0

# Authentication & Security
passlib[argon2]>=1.7.4
argon2-cffi>=21.0.0
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0

# Rate Limiting (for 1000-3000+ concurrent users)
slowapi>=0.1.9

# HTTP Client & File Handling
httpx>=0.25.0
python-multipart>=0.0.6

# PDF Processing
pdfplumber>=0.10.0

# Additional utilities (if needed by extension/frontend integration)
python-dateutil>=2.8.0

# Testing & Development (Optional but recommended)
pytest>=7.4.0
pytest-asyncio>=0.21.0
